+++ src/services/index.ts
@@ -1,3 +1,4 @@
 export * from './configService';
 export * from './tasksService';
 export * from './resouceService';
+export * from './scriptService';