+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased][Unreleased]
 
+## [0.4.0][0.4.0]
+
 ### Added
 
 - Add launcher view
@@ -127,7 +129,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add auto refresh
 - Add runBash in `utils`
 
-[unreleased]: https://github.com/XingYuSSS/slurm--/compare/v0.3.1...HEAD
+[unreleased]: https://github.com/XingYuSSS/slurm--/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/XingYuSSS/slurm--/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/XingYuSSS/slurm--/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/XingYuSSS/slurm--/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/XingYuSSS/slurm--/compare/v0.1.2...v0.2.0